Added a skiplock on the asset cleanup
And moved the asset ID to the front of the logging statements.  That made tracking parallel task operations much easier.

The skiplock is designed to help prevent multiple calls trying to do a cleanup operation so that the DB doesn't get overcleaned.  Overcleaning can still happen, but it's not near as dramatic as before the skiplock.