Made asset IDs known ASAP, even if not on disk yet.
This solved the double PUT problem, where if two PUTs for the same asset ID are sent too fast it could cause a conflict.

GET will spinwait until the actual data is available on disk, blocking the client and hopefully preventing the whole "I PUT it, so why am I getting a Not Found?!" problem.