Attempting fix for #6: wcache expansion works, but shrinking is bypassed
UNTESTED: Can't get NUnit to behave on OSX, will test later.

Expanding the file is easy: just add more NUL bytes to the end.  Shrinking the file, not so easy, especially if there are occupied records.  So thus I punted: expanding should work, but shrinking doesn't happen: instead it just uses the size from the file.

A TODO entry was added to the code to mark the interest for support of shrinking the write cache.